requirements.txt
QuantLib
pandas
pydantic
python-dotenv
python-dateutil
uvicorn
fastapi
fastapi_azure_auth